+++ CHANGELOG.md
@@ -4,6 +4,11 @@
 
 * Update dependencies for security fixes
 
+## v0.11.0
+
+* handle new Openstack "attaching" state
+* fix issues with unmount mutex (could lead to deadlocks) in version v0.10.0
+
 ## v0.10.0
 
 * error handling at mount: umount/cleanup if mount fails